x86_64: Use the i386 arch_options_t on x86_64
In preparation for having a single crashdump implementation
on i386 and x86_64 start sharing arch_options_t between
i386 and x86_64, and don't make arch_options static on
x86_64.

i386 and x86_64 have been using identitcal defitions of
arch_options for a while this change now reflects that
practical reality in the code.

Signed-off-by: Eric W. Biederman <ebiederm@xmission.com>
Signed-off-by: Simon Horman <horms@verge.net.au>