kexec elf: Clean up error messages build_mem_phdrs
Don't print uncondintionally print an error message about
KCORE_ELF_HEADERS_SIZE being wrong when the elf headers
are truncated when build_mem_phdrs fails because there
are other reasons we could be truncated (like an
actual damaged elf file).

Instead only print an error message when either probe_debug
is true or we are skipping the total file size checks for
some reason.  When we do print an error message limit the
message to the fact that we are truncated and how many
bytes we have and how many bytes we need.  The context
of the failure should be able to provide the rest of
the information.

For now add a comment in crashdump-x86_64.c to direct people
to KCORE_ELF_HEADERS_SIZE so that if someone finds the
code things can be adjusted.

Signed-off-by: Eric W. Biederman <ebiederm@xmission.com>
Signed-off-by: Simon Horman <horms@verge.net.au>