x86 bzImage: Optimize the panic kernel memory layout.
Load the setup sectors above the kernel in the panic case,
to avoid wasting memory.

Linux kernels frequently have a severe alignment requirment
that can be as much as 16M.  The start of the reserved
crash kernel region is 16M aligned.  The setup sectors
are tiny 20k? and when we are not running the real mode code
they can live anywhere.  Move the setup sectors from
lower address than the kernel to higher address than the kernel
avoids introducing a bad alignment.  Allowing the linux kernel
to load lower.  Loading the linux kernel lower in memory ensures
memory is used more efficiently.

Signed-off-by: Eric W. Biederman <ebiederm@xmission.com>
Signed-off-by: Simon Horman <horms@verge.net.au>