get the backup area dynamically
Currently we hard-coded the first 640K area as backup area,
however, this is not correct on some system which has reserved
memory area in the first 640K:

BIOS-e820: 0000000000010000 - 0000000000097000 (usable)
BIOS-e820: 0000000000097000 - 00000000000a0000 (reserved)

on such system we still mark all the first 640K as usable
in capture kernel, this will cause kernel panic.

The solution, pointed by Vivek, is that we can get the backup
area dynamically by reading /proc/iomem.

The reporter has tested this patch and it fixes the problem.

Signed-off-by: WANG Cong <amwang@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Simon Horman <horms@verge.net.au>