Add generic debug option
Currently the debugging code is under #ifdef DEBUG, which
means when we want to debug, we have to re-compile the source
code with -DDEBUG. This is not convenient, we want to have
a generic --debug option so that we can enable debugging code
without re-compiling.

This patch moves the arch-specific --debug to generic place
and moves code under #ifdef DEBUG to --debug on x86.

BTW, the size of kexec binary increases very little after this patch.

Signed-off-by: Cong Wang <xiyou.wangcong@gmail.com>
Signed-off-by: Simon Horman <horms@verge.net.au>