Load bzImages smaller than 32 KiB
Allow bzImages smaller than 32KiB to be kexec'ed.

The current code will fail to load a bzImage smaller than 32768 bytes (sizeof
struct x86_linux_header), but the 'memdisk' program that comes with syslinux is
only about 26 KiB.  This patch changes the minimum size to 1024 bytes (2
sectors), which appears to be the limit that syslinux enforces.

Removed the "tail" field of struct x86_linux_header because it doesn't seem to
actually be used (is there a reason for it?).

Also, note that bzImage_probe() was incorrectly using `sizeof (header)', even
though header is a pointer.

Signed-off-by: Eric Biggers <ebiggers3@gmail.com>
Signed-off-by: Simon Horman <horms@verge.net.au>