fs2dt: Add a generic copy of fs2dt
The motivation for this is to remove duplicated code by sharing the fs2dt
between different architectures.

The code added by this patch is very close to the code currently used by
ppc64, and thus migrating that architecture should not be difficult.  The
32bit powerpc code is a little different and thus more care is needed when
migrating that architecture to this code.

Unfortunately I do not have any powerpc equipment available to test this
code.

Signed-off-by: Simon Horman <horms@verge.net.au>