kexec: Teach configure to find the strip binary.
For some reason my version of the Makefile generated by configure
included the line STRIP=strp.  Rerunning configure from a fresh slate
did not regenerate that line so I don't know how it got there.  So add
the code to Makefile.in and configure.ac to autodetect the strip binary.

This is needed so that we can remove from purgatory all of the
relocations to sections that are not needed at runtime, by stripping
out those sections.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
[horms@verge.net.au: Omitted white-space only change to
                     purgatory/arch/x86_64/Makefile]
Signed-off-by: Simon Horman <horms@verge.net.au>