kexec x86: drop truncation warning for crash kernel
On kexec set-up of a crash kernel on a very large memory machine we
sometimes see the worrisome warning:
  Too many memory ranges, truncating...
meaning that the total count of e820 ram, reserved and ACPI spaces is over 128.

Per the comment in do_bzImage_load():
                /* If using bzImage for capture kernel, then we will not be
                 * executing real mode code. setup segment can be loaded
                 * anywhere as we will be just reading command line.
                 */
So if I understand, the e820 table built here and added to the 'real_mode'
area is not going to be used.
So the warning message can be omitted.

Signed-off-by: Cliff Wickman <cpw@sgi.com>
[ horms@verge.net.au manually applied due to conflict ]
Signed-off-by: Simon Horman <horms@verge.net.au>